introduce message preview list component
